Start refactoring layout mouse handles to move functionality to a generic base class
